Extract BannerAdmin component
